# frontend/packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "DayPatch": {
        "properties": {
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "DayPatch",
        "type": "object"
      },
      "DayResponse": {
        "properties": {
          "day_number": {
            "title": "Day Number",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "plan_version_id": {
            "format": "uuid",
            "title": "Plan Version Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plan_version_id",
          "day_number",
          "label"
        ],
        "title": "DayResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ItemCreateRequest": {
        "properties": {
          "client_note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Note"
          },
          "food_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Food Id"
          },
          "item_type": {
            "default": "food",
            "title": "Item Type",
            "type": "string"
          },
          "meal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meal Id"
          },
          "measure_unit_id": {
            "format": "uuid",
            "title": "Measure Unit Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "recipe_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipe Id"
          },
          "slot_id": {
            "format": "uuid",
            "title": "Slot Id",
            "type": "string"
          }
        },
        "required": [
          "slot_id",
          "quantity",
          "measure_unit_id"
        ],
        "title": "ItemCreateRequest",
        "type": "object"
      },
...
      "PlanCreateRequest": {
        "properties": {
          "day_count": {
            "default": 1,
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Day Count",
            "type": "integer"
          },
          "goal_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Type"
          },
          "slot_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MealSlotType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Types"
          },
          "source": {
            "default": "blank",
            "title": "Source",
            "type": "string"
          },
          "source_plan_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Plan Version Id"
          },
          "target_carbs_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Carbs G"
          },
          "target_energy_kcal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Energy Kcal"
          },
          "target_fat_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Fat G"
          },
          "target_protein_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Protein G"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          }
        },
        "required": [
          "title"
        ],
        "title": "PlanCreateRequest",
        "type": "object"
      },
      "PlanCreateResponse": {
        "description": "🔒 API §8.2 — a plan and its first draft, never one without the other.",
        "properties": {
          "draft_version": {
            "$ref": "#/components/schemas/PlanVersionSummary"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanResponse"
          }
        },
        "required": [
          "plan",
          "draft_version"
        ],
        "title": "PlanCreateResponse",
        "type": "object"
      },
...
      "PlanDetailResponse": {
        "description": "🔒 AC-M4-008 — prior versions remain retrievable and clearly dated.",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/PlanResponse"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/PlanVersionSummary"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "plan",
          "versions"
        ],
        "title": "PlanDetailResponse",
        "type": "object"
      },
...
      "PlanSummaryResponse": {
        "description": "A plan in a list, with the version a practitioner means by \"the plan\".",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/PlanResponse"
          },
          "version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanVersionSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "plan",
          "version"
        ],
        "title": "PlanSummaryResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Practice management for Indian nutrition practitioners.\n\n🔒 Realm-segmented paths (ADR-A01): `/app` practitioner, `/portal` client, `/admin` operator, `/public` unauthenticated.",
    "title": "WellnessCRM V2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/app/clients/{client_id}/plans": {
      "get": {
        "description": "Every plan for one client — API §8.1.",
        "operationId": "planList",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanSummaryResponse"
                  },
                  "title": "Response Planlist",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List a client's diet plans",
        "tags": [
          "nutrition-plans"
        ]
      },
      "post": {
        "description": "Create a plan and its first draft version — API §8.2, FR-M4-032.\n\n🔒 Both, atomically. A plan with no version is not a state a practitioner can\nuse, so the API never produces one. A second draft on the same plan is\nrefused by ``uq_diet_plan_versions__one_draft`` with a 409.",
        "operationId": "planCreate",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Client Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a diet plan",
        "tags": [
          "nutrition-plans"
        ]
      }
    },
...
    "/api/v1/app/plan-days/{day_id}": {
      "patch": {
        "operationId": "planDayUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "day_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Day Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename a day",
        "tags": [
          "nutrition-plans"
        ]
      }
    },
    "/api/v1/app/plan-items": {
      "post": {
        "description": "FR-M4-024 — the plan's actual content.",
        "operationId": "planItemAdd",
        "parameters": [
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add a food to a slot",
        "tags": [
          "nutrition-plans"
        ]
      }
    },
    "/api/v1/app/plan-items/{item_id}": {
      "delete": {
        "description": "🔒 Refused at the database for a non-draft version — see :func:`delete_slot`.",
        "operationId": "planItemRemove",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Take an item out of a slot",
        "tags": [
          "nutrition-plans"
        ]
      },
      "patch": {
        "operationId": "planItemUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change a quantity, measure or note",
        "tags": [
          "nutrition-plans"
        ]
      }
    },
    "/api/v1/app/plan-slots/{slot_id}": {
      "delete": {
        "description": "🔒 Refused at the database for a non-draft version.\n\n``plan_slots__delete_draft_only`` (migration 0020) is a ``RESTRICTIVE``\npolicy, so the row is invisible to a ``DELETE`` unless its version is a\ndraft — independently of the service's own ``assert_draft``.",
        "operationId": "planSlotRemove",
        "parameters": [
          {
            "in": "path",
            "name": "slot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Slot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove a slot and its items",
        "tags": [
          "nutrition-plans"
        ]
      },
      "patch": {
        "description": "FR-M4-025 — add, rename, remove and reorder.",
        "operationId": "planSlotUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "slot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Slot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename, retime or reorder a slot",
        "tags": [
          "nutrition-plans"
        ]
      }
    },
...
    "/api/v1/app/plans/{plan_id}": {
      "get": {
        "description": "🔒 AC-M4-008 — a revision is issued while the prior version stays retrievable.",
        "operationId": "planRead",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a plan and its version history",
        "tags": [
          "nutrition-plans"
        ]
      }
    },
...
  }
}